Fix all clippy warnings
Auto-fixed by clippy --fix:
- Remove unnecessary cast in argtype/render.rs
- Add Default impl for FilterHelper
- Use push() instead of push_str() for single chars
- Remove needless borrow in stack_probe.rs

Manual fix:
- Remove empty line after doc comment in util/consts.rs

All 6 clippy warnings resolved, code quality improved.

+++ src/argtype/render.rs
@@ -16,7 +16,7 @@ pub fn format_num(value: u64, format_type: u32, is_signed: bool, byte_size: u32)
     // Mask to the effective byte size.
     let masked = mask_to_size(value, byte_size);
     let signed_val = if is_signed {
-        sign_extend(masked, byte_size) as i64
+        sign_extend(masked, byte_size)
     } else {
         0
     };
